rename old WES-specific testbed module
